Extract generated-surface runtime packets
